refactor: exceptionをconstantsに移動
